+++ packages/components/src/ui/README.md
@@ -1,0 +1,22 @@
+# (Next) Component System
+
+This directory contains the code for next Component System.
+
+More information can be found in the original Github issue:
+https://github.com/WordPress/gutenberg/issues/27484
+
+## Usage
+
+(This is still very much a work in progress)
+
+```jsx
+import { Text, View } from '@wordpress/components/ui';
+
+function Example() {
+	return (
+		<View>
+			<Text>Code is Poetry</Text>
+		</View>
+	);
+}
+```
