--- packages/components/src/ui/control-label/control-label.js
@@ -1,13 +1,0 @@
-/**
- * Internal dependencies
- */
-import { createComponent } from '../utils';
-import { useControlLabel } from './use-control-label';
-
-const ControlLabel = createComponent( {
-	as: 'label',
-	useHook: useControlLabel,
-	name: 'ControlLabel',
-} );
-
-export default ControlLabel;
