+++ packages/components/src/ui/font-size-control/README.md
@@ -1,0 +1,5 @@
+# FontSizeControl
+
+`FontSizeControl` is a tool that renders a UI that allows users to select a font size.
+
+For more information, see the [`FontSizePicker`](../../font-size-picker).
