+++ packages/components/src/ui/form-group/README.md
@@ -5,10 +5,7 @@
 ## Usage
 
 ```jsx
-import {
-	__experimentalFormGroup as FormGroup,
-	__experimentalTextInput as TextInput,
-} from '@wordpress/components';
+import { FormGroup, TextInput } from '@wordpress/components/ui';
 
 function Example() {
 	return (
