+++ packages/components/src/ui/h-stack/component.js
@@ -4,6 +4,32 @@
 import { createComponent } from '../utils';
 import { useHStack } from './hook';
 
+/**
+ *`HStack` (Horizontal Stack) arranges child elements in a horizontal line.
+ *
+ * `HStack` can render anything inside.
+ *
+ * @example
+ * ```jsx
+ * import { HStack, Text, View } from `@wp-g2/components/ui`;
+ *
+ * function Example() {
+ * 	return (
+ * 		<HStack>
+ * 			<View>
+ * 				<Text>Code</Text>
+ * 			</View>
+ * 			<View>
+ * 				<Text>is</Text>
+ * 			</View>
+ * 			<View>
+ * 				<Text>Poetry</Text>
+ * 			</View>
+ * 		</HStack>
+ * 	);
+ * }
+ * ```
+ */
 export default createComponent( {
 	as: 'div',
 	useHook: useHStack,
