+++ packages/components/src/ui/h-stack/types.ts
@@ -38,28 +38,6 @@ export type Props = Omit< FlexProps, 'align' | 'gap' > & {
 	 * * `stretch`: Stretches content to the edges of the container.
 	 *
 	 * @default 'edge'
-	 *
-	 * @example
-	 *```jsx
-	 * import { HStack, Text, View } from `@wp-g2/components`
-	 * import { ui } from `@wp-g2/styles`
-	 *
-	 * function Example() {
-	 *   return (
-	 *     <HStack alignment="center">
-	 *       <View css={[ui.background.blue]}>
-	 *         <Text>Ana</Text>
-	 *       </View>
-	 *       <View css={[ui.background.blue]}>
-	 *         <Text>Elsa</Text>
-	 *       </View>
-	 *       <View css={[ui.background.blue]}>
-	 *         <Text>Olaf</Text>
-	 *       </View>
-	 *     </HStack>
-	 *   );
-	 * }
-	 *```
 	 */
 	alignment?: HStackAlignment | CSSProperties[ 'alignItems' ];
 	/**
@@ -67,57 +45,6 @@ export type Props = Omit< FlexProps, 'align' | 'gap' > & {
 	 * The value of `spacing` works as a multiplier to the library's grid system (base of `4px`).
 	 *
 	 * @default 2
-	 *
-	 * @example
-	 * ```jsx
-	 * import { HStack, Text, View } from `@wp-g2/components`
-	 * import { ui } from `@wp-g2/styles`
-	 *
-	 * function Example() {
-	 *   return (
-	 *     <HStack alignment="center" spacing={8}>
-	 *       <View css={[ui.background.blue]}>
-	 *         <Text>Ana</Text>
-	 *       </View>
-	 *       <View css={[ui.background.blue]}>
-	 *         <Text>Elsa</Text>
-	 *       </View>
-	 *       <View css={[ui.background.blue]}>
-	 *         <Text>Olaf</Text>
-	 *       </View>
-	 *     </HStack>
-	 *   );
-	 * }
-	 *```
 	 */
 	spacing?: CSSProperties[ 'width' ];
 };
-
-/**
- * `HStack` (Horizontal Stack) arranges child elements in a horizontal line.
- *
- * @remarks
- * `HStack` can render anything inside.
- *
- * @example
- * ```jsx
- * import { HStack, Text, View } from `@wp-g2/components`
- * import { ui } from `@wp-g2/styles`
- *
- * function Example() {
- *   return (
- *     <HStack>
- *       <View css={[ui.background.blue]}>
- *         <Text>Ana</Text>
- *       </View>
- *       <View css={[ui.background.blue]}>
- *         <Text>Elsa</Text>
- *       </View>
- *       <View css={[ui.background.blue]}>
- *         <Text>Olaf</Text>
- *       </View>
- *     </HStack>
- *   );
- * }
- * ```
- */
