+++ packages/components/src/ui/surface/component.js
@@ -4,6 +4,23 @@
 import { createComponent } from '../utils';
 import { useSurface } from './hook';
 
+/**
+ * `Surface` is a core component that renders a primary background color.
+ *
+ * In the example below, notice how the `Surface` renders in white (or dark gray if in dark mode).
+ *
+ * ```jsx
+ * import { Surface, Text } from `@wp-g2/components/ui`;
+ *
+ * function Example() {
+ * 	return (
+ * 		<Surface>
+ * 			<Text>Code is Poetry</Text>
+ * 		</Surface>
+ * 	);
+ * }
+ * ```
+ */
 export default createComponent( {
 	as: 'div',
 	useHook: useSurface,
