+++ packages/components/src/ui/text/types.ts
@@ -23,36 +23,10 @@ type TextWeight = 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900;
 export interface Props extends TruncateProps {
 	/**
 	 * Adjusts the text alignment.
-	 *
-	 * @example
-	 * ```jsx
-	 * import { Text } from `@wordpress/components/ui`
-	 *
-	 * function Example() {
-	 * 	return (
-	 * 		<Text align="center" isBlock>
-	 * 			Where the north wind meets the sea...
-	 * 		</Text>
-	 * 	)
-	 * }
-	 *```
 	 */
 	align?: CSSProperties[ 'textAlign' ];
 	/**
 	 * Automatically calculate the appropriate line-height value for contents that render text and Control elements (e.g. `TextInput`).
-	 *
-	 * @example
-	 * ```jsx
-	 * import { Text, TextInput } from `@wordpress/components/ui`
-	 *
-	 * function Example() {
-	 * 	return (
-	 * 		<Text adjustLineHeightForInnerControls>
-	 * 			Where the north wind meets the <TextInput value="sea..." />
-	 * 		</Text>
-	 * 	)
-	 * }
-	 *```
 	 */
 	adjustLineHeightForInnerControls?: TextAdjustLineHeightForInnerControls;
 	/**
@@ -91,57 +65,14 @@ export interface Props extends TruncateProps {
 	lineHeight?: CSSProperties[ 'lineHeight' ];
 	/**
 	 * The `Text` color can be adapted to a background color for optimal readability. `optimizeReadabilityFor` can accept CSS variables, in addition to standard CSS color values (e.g. Hex, RGB, HSL, etc...).
-	 *
-	 * @example
-	 * ```jsx
-	 * import { Text, View } from `@wordpress/components/ui`
-	 *
-	 * function Example() {
-	 *   const backgroundColor = "blue"
-	 *
-	 *   return (
-	 *     <View css={{ backgroundColor }}>
-	 *       <Text optimizeReadabilityFor={backgroundColor}>
-	 *         Where the north wind meets the sea, there's a river full of memory.
-	 *       </Text>
-	 *     </View>
-	 *   )
-	 * }
-	 * ```
 	 */
 	optimizeReadabilityFor?: CSSProperties[ 'color' ];
 	/**
 	 * Adjusts text size based on the typography system. `Text` can render a wide range of font sizes, which are automatically calculated and adapted to the typography system. The `size` value can be a system preset, a `number`, or a custom unit value (`string`) such as `30em`.
-	 *
-	 * @example
-	 * ```jsx
-	 * import { Text } from `@wordpress/components/ui`
-	 *
-	 * function Example() {
-	 *   return <Text size="largeTitle">Where the north wind meets the sea...</Text>
-	 * }
-	 * ```
 	 */
 	size?: CSSProperties[ 'fontSize' ] | TextSize;
 	/**
 	 * Enables text truncation. When `truncate` is set,we are able to truncate the long text in a variety of ways.
-	 *
-	 * @example
-	 *
-	 * ```jsx
-	 * import { Text } from `@wordpress/components/ui`
-	 *
-	 * function Example() {
-	 *   return (
-	 *     <Text truncate>
-	 *       Where the north wind meets the sea, there's a river full of memory. Sleep,
-	 *       my darling, safe and sound, for in this river all is found. In her waters,
-	 *       deep and true, lay the answers and a path for you. Dive down deep into her
-	 *       sound, but not too far or you'll be drowned
-	 *     </Text>
-	 *   )
-	 * }
-	 * ```
 	 */
 	truncate?: boolean;
 	/**
@@ -150,15 +81,6 @@ export interface Props extends TruncateProps {
 	upperCase?: boolean;
 	/**
 	 * Adjusts style variation of the text.
-	 *
-	 * @example
-	 * ```jsx
-	 * import { Text } from `@wordpress/components/ui`
-	 *
-	 * function Example() {
-	 *   return <Text variant="muted">Where the north wind meets the sea...</Text>
-	 * }
-	 * ```
 	 */
 	variant?: TextVariant;
 	/**
@@ -171,22 +93,6 @@ export interface Props extends TruncateProps {
 	letterSpacing?: CSSProperties[ 'letterSpacing' ];
 	/**
 	 * Letters or words within `Text` can be highlighted using `highlightWords`.
-	 *
-	 * @example
-	 * ```jsx
-	 * import { Text } from `@wordpress/components/ui`
-	 *
-	 * function Example() {
-	 *   return (
-	 *     <Text highlightWords={["the"]}>
-	 *       Where the north wind meets the sea, there's a river full of memory. Sleep,
-	 *       my darling, safe and sound, for in this river all is found. In her waters,
-	 *       deep and true, lay the answers and a path for you. Dive down deep into her
-	 *       sound, but not too far or you'll be drowned
-	 *     </Text>
-	 *   )
-	 * }
-	 * ```
 	 */
 	highlightWords?: string[];
 }
