+++ packages/components/src/ui/truncate/README.md
@@ -10,10 +10,9 @@ import { Truncate } from '@wordpress/components/ui';
 function Example() {
 	return (
 		<Truncate>
-			Where the north wind meets the sea, there's a river full of memory.
-			Sleep, my darling, safe and sound, for in this river all is found.
-			In her waters, deep and true, lay the answers and a path for you.
-			Dive down deep into her sound, but not too far or you'll be drowned
+			Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc ex
+			neque, vulputate a diam et, luctus convallis lacus. Vestibulum ac
+			mollis mi. Morbi id elementum massa.
 		</Truncate>
 	);
 }
