+++ packages/components/src/ui/truncate/component.js
@@ -5,21 +5,23 @@ import { createComponent } from '../utils';
 import useTruncate from './hook';
 
 /**
- * `Truncate` is a typography primitive that trims text content. For almost all cases, it is recommended that `Text`, `Heading`, or `Subheading` is used to render text content. However, `Truncate` is available for custom implementations.
+ * `Truncate` is a typography primitive that trims text content.
+ * For almost all cases, it is recommended that `Text`, `Heading`, or
+ * `Subheading` is used to render text content. However,`Truncate` is
+ * available for custom implementations.
  *
  * @example
  * ```jsx
- * import { Truncate } from `@wp-g2/components`
+ * import { Truncate } from `@wordpress/components/ui`;
  *
  * function Example() {
- *   return (
- *     <Truncate>
- *       Where the north wind meets the sea, there's a river full of memory. Sleep,
- *       my darling, safe and sound, for in this river all is found. In her waters,
- *       deep and true, lay the answers and a path for you. Dive down deep into her
- *       sound, but not too far or you'll be drowned
- *     </Truncate>
- *   );
+ * 	return (
+ * 		<Truncate>
+ * 			Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc ex
+ * 			neque, vulputate a diam et, luctus convallis lacus. Vestibulum ac
+ * 			mollis mi. Morbi id elementum massa.
+ * 		</Truncate>
+ * 	);
  * }
  * ```
  */
