+++ packages/components/src/ui/truncate/types.ts
@@ -16,22 +16,6 @@ export interface Props {
 	 * * `tail`: Trims content at the end. Requires a `limit`.
 	 *
 	 * @default 'auto'
-	 *
-	 * @example
-	 * ```jsx
-	 * import { Truncate } from `@wp-g2/components`
-	 *
-	 * function Example() {
-	 * 	return (
-	 * 		<Truncate ellipsizeMode="middle" limit={40}>
-	 * 			Where the north wind meets the sea, there's a river full of memory. Sleep,
-	 * 			my darling, safe and sound, for in this river all is found. In her waters,
-	 * 			deep and true, lay the answers and a path for you. Dive down deep into her
-	 * 			sound, but not too far or you'll be drowned
-	 * 		</Truncate>
-	 * 	)
-	 * }
-	 * ```
 	 */
 	ellipsizeMode?: TruncateEllisizeMode;
 	/**
@@ -42,22 +26,6 @@ export interface Props {
 	limit?: number;
 	/**
 	 * Clamps the text content to the specifiec `numberOfLines`, adding the `ellipsis` at the end.
-	 *
-	 * @example
-	 * ```jsx
-	 * import { Truncate } from `@wp-g2/components`
-	 *
-	 * function Example() {
-	 * 	return (
-	 * 		<Truncate numberOfLines={2}>
-	 * 			Where the north wind meets the sea, there's a river full of memory. Sleep,
-	 * 			my darling, safe and sound, for in this river all is found. In her waters,
-	 * 			deep and true, lay the answers and a path for you. Dive down deep into her
-	 * 			sound, but not too far or you'll be drowned
-	 * 		</Truncate>
-	 * 	)
-	 * }
-	 * ```
 	 */
 	numberOfLines?: number;
 }
