+++ packages/components/src/ui/v-stack/README.md
@@ -6,21 +6,20 @@
 
 `VStack` can render anything inside.
 
-```jsx live
-import { VStack, Text, View } from '@wp-g2/components';
-import { ui } from '@wp-g2/styles';
+```jsx
+import { VStack, Text, View } from '@wp-g2/components/ui';
 
 function Example() {
 	return (
-		<VStack css={[ui.frame.height(200)]}>
-			<View css={[ui.background.blue]}>
-				<Text>Ana</Text>
+		<VStack>
+			<View>
+				<Text>Code</Text>
 			</View>
-			<View css={[ui.background.blue]}>
-				<Text>Elsa</Text>
+			<View>
+				<Text>is</Text>
 			</View>
-			<View css={[ui.background.blue]}>
-				<Text>Olaf</Text>
+			<View>
+				<Text>Poetry</Text>
 			</View>
 		</VStack>
 	);
@@ -86,14 +85,14 @@ When a `Spacer` is used within an `VStack`, the `Spacer` adaptively expands to t
 ```jsx
 function Example() {
 	return (
-		<VStack css={[ui.frame.height(200)]}>
-			<View css={[ui.background.blue]}>
+		<VStack css={ [ ui.frame.height( 200 ) ] }>
+			<View css={ [ ui.background.blue ] }>
 				<Text>Ana</Text>
 			</View>
-			<Spacer css={[ui.background.blue]}>
+			<Spacer css={ [ ui.background.blue ] }>
 				<Text>Elsa</Text>
 			</Spacer>
-			<View css={[ui.background.blue]}>
+			<View css={ [ ui.background.blue ] }>
 				<Text>Olaf</Text>
 			</View>
 		</VStack>
@@ -103,18 +102,18 @@ function Example() {
 
 `Spacer` can also be used in-between items to push them apart.
 
-```jsx live
+```jsx
 function Example() {
 	return (
-		<VStack css={[ui.frame.height(200)]}>
-			<View css={[ui.background.blue]}>
+		<VStack css={ [ ui.frame.height( 200 ) ] }>
+			<View css={ [ ui.background.blue ] }>
 				<Text>Ana</Text>
 			</View>
 			<Spacer />
-			<View css={[ui.background.blue]}>
+			<View css={ [ ui.background.blue ] }>
 				<Text>Elsa</Text>
 			</View>
-			<View css={[ui.background.blue]}>
+			<View css={ [ ui.background.blue ] }>
 				<Text>Olaf</Text>
 			</View>
 		</VStack>
