+++ packages/components/src/ui/view/component.js
@@ -1,0 +1,26 @@
+/**
+ * External dependencies
+ */
+import { BaseView } from '@wp-g2/styles';
+
+/**
+ * `View` is a core component that renders everything in the library.
+ * It is the principle component in the entire library.
+ *
+ * @example
+ * ```jsx
+ * import { View } from `@wordpress/components/ui`;
+ *
+ * function Example() {
+ * 	return (
+ * 		<View>
+ * 			 Code is Poetry
+ * 		</View>
+ * 	);
+ * }
+ * ```
+ */
+const View = BaseView;
+View.displayName = 'View';
+
+export default View;
