+++ packages/components/src/ui/view/index.js
@@ -1,1 +1,1 @@
-export { default as View } from './view';
+export { default as View } from './component';
