--- packages/components/src/ui/view/view.js
@@ -1,9 +1,0 @@
-/**
- * External dependencies
- */
-import { BaseView } from '@wp-g2/styles';
-
-const View = BaseView;
-View.displayName = 'View';
-
-export default View;
